+++ src/Network/Network/VirtualRouter/VirtualRouterParameterSetNames.cs
@@ -16,8 +16,7 @@ namespace Microsoft.Azure.Commands.Network
 {
     internal static class VirtualRouterParameterSetNames
     {
-        internal const string ByHostedGateway = "HostedGatewayParameterSet";
-        internal const string ByHostedGatewayId = "HostedGatewayIdParameterSet";
+        internal const string ByVirtualRouterSubscriptionId = "VirtualRouterSubscriptionIdParameterSet";
         internal const string ByVirtualRouterResourceId = "VirtualRouterResourceIdParameterSet";
         internal const string ByVirtualRouterInputObject = "VirtualRouterInputObjectParameterSet";
         internal const string ByVirtualRouterName = "VirtualRouterNameParameterSet";