+++ src/Sql/Sql/Database/Model/AzureSqlDatabaseModel.cs
@@ -181,6 +181,11 @@ namespace Microsoft.Azure.Commands.Sql.Database.Model
         public int? ReadReplicaCount { get; set; }
 
         /// <summary>
+        /// Gets or sets the backup storage redundancy for the database
+        /// </summary>
+        public string BackupStorageRedundancy { get; set; }
+
+        /// <summary>
         /// Construct AzureSqlDatabaseModel
         /// </summary>
         public AzureSqlDatabaseModel()
@@ -233,6 +238,7 @@ namespace Microsoft.Azure.Commands.Sql.Database.Model
             AutoPauseDelayInMinutes = null;
             MinimumCapacity = null;
             ReadReplicaCount = null;
+            BackupStorageRedundancy= null;
         }
 
         /// <summary>
@@ -284,6 +290,27 @@ namespace Microsoft.Azure.Commands.Sql.Database.Model
             AutoPauseDelayInMinutes = database.AutoPauseDelay;
             MinimumCapacity = database.MinCapacity;
             ReadReplicaCount = database.ReadReplicaCount;
+            BackupStorageRedundancy = MapInternalBackupStorageRedundancyToExternal(database.StorageAccountType);
+        }
+
+        /// <summary>
+        /// Map internal BackupStorageRedundancy value (GRS/LRS/ZRS) to external (Geo/Local/Zone)
+        /// </summary>
+        /// <param name="backupStorageRedundancy">Backup storage redundancy</param>
+        /// <returns>internal backupStorageRedundancy</returns>
+        private static string MapInternalBackupStorageRedundancyToExternal(string backupStorageRedundancy)
+        {
+            switch (backupStorageRedundancy)
+            {
+                case "GRS":
+                    return "Geo";
+                case "LRS":
+                    return "Local";
+                case "ZRS":
+                    return "Zone";
+                default:
+                    return null;
+            }
         }
     }
 }