+++ tools/GenerateHelp.ps1
@@ -99,6 +99,8 @@ $NewModules = @("Az.AppConfiguration",
                 "Az.MySql",
                 "Az.Portal",
                 "Az.PostgreSql",
+                "Az.ResourceMover",
+                "Az.SpringCloud",
                 "Az.TimeSeriesInsights"
                 )
 if ($GenerateMamlHelp)