+++ src/CognitiveServices/CognitiveServices/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.6.0
+* Showed additional legal terms for certain APIs.
+
 ## Version 1.5.1
 * Used `Deny` specifically as NetworkRules default action.