+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.10.1
+* Updated ADF .Net SDK version to 4.11.0
+
+## Version 1.10.0
 * Added missing properties to PSPipelineRun class.
 
 ## Version 1.9.0